refactor: split commands.rs (820 LOC) into per-command modules (#180)
Split monolithic commands.rs into 9 focused modules:
- server.rs (183 LOC) — HTTP routing for server mode
- remember.rs (59 LOC) — store new memories
- recall.rs (57 LOC) — semantic + keyword search
- list.rs (54 LOC) — list, get, stats
- forget.rs (96 LOC) — delete (single, tag, cold, all)
- tags.rs (74 LOC) — tag CRUD
- namespace.rs (56 LOC) — namespace management
- aging.rs (89 LOC) — aging status/preview/cleanup
- maintenance.rs (230 LOC) — doctor, verify, repair, prune, consolidate, export, import, checksums

Zero behavioral changes — pure reorganization.
