+++ crates/uteke-cli/src/commands/recall.rs
@@ -1,0 +1,57 @@
+//! Recall and Search commands — semantic and keyword search.
+
+use crate::output;
+use crate::Cli;
+use uteke_core::Uteke;
+
+pub(crate) fn run_recall(
+    cli: &Cli,
+    uteke: &Uteke,
+    ns: Option<&str>,
+    query: &str,
+    limit: usize,
+    tags: &[String],
+) -> Result<(), String> {
+    tracing::info!("Recalling: {query} (limit: {limit})");
+    let tag_refs: Vec<&str> = tags.iter().map(|s| s.as_str()).collect();
+    let tags_filter = if tag_refs.is_empty() {
+        None
+    } else {
+        Some(tag_refs.as_slice())
+    };
+    let results = uteke
+        .recall(query, limit, tags_filter, ns)
+        .map_err(|e| format!("Failed to recall: {e}"))?;
+    if cli.json {
+        output::print_json(&results);
+    } else {
+        output::print_recall_human(&results);
+    }
+    Ok(())
+}
+
+pub(crate) fn run_search(
+    cli: &Cli,
+    uteke: &Uteke,
+    ns: Option<&str>,
+    query: &str,
+    limit: usize,
+    tags: &[String],
+) -> Result<(), String> {
+    tracing::info!("Searching: {query} (limit: {limit})");
+    let tag_refs: Vec<&str> = tags.iter().map(|s| s.as_str()).collect();
+    let tags_filter = if tag_refs.is_empty() {
+        None
+    } else {
+        Some(tag_refs.as_slice())
+    };
+    let results = uteke
+        .search(query, limit, tags_filter, ns)
+        .map_err(|e| format!("Failed to search: {e}"))?;
+    if cli.json {
+        output::print_json(&results);
+    } else {
+        output::print_search_human(&results);
+    }
+    Ok(())
+}
